[FIX] estate: change `state` selection T#69469
In the model `estate.property.state` its `state` property was set to
`("new", "offer", "received", "accepted", "sold", "canceled")` as I
misreaad what the states nedded to be.

The correct posible states for a property are of `"("new",
"offer_received", "offer_accepted", "sold", "canceled")`, as it was
originally requested in the section "Getting Started 6".