[IMP] estate: implement basic views T#69469
In the previous `[IMP]` commit I added the basic files necessary to have
a basic (default) view of the fields of our model. In this commit I
implement the `List`, `Form` and `Search` views.

Every view was implemented as to what is specified in the section
"Getting Started 7" of the technical training. These views were
implemented to improve the UI/UX of the module.

In the `List` view several tags are shown. In the `Form` view the layout
and visibility of fields of the `estate.property` fields were modified.
Finally, in the `Search` "view" `Group By` by postcode and `Filter` by
`'state' were added.