[IMP] estate: add tag, type and offers T#69469
Add models, views and relationships for `estate.property.type`,
`estate.property.tag` and `estate.property.offer`.

Everything was implemented as specified in the section "Getting Started 8".

Property type was added to give more information to future customer about the
kind of building some property is. Offers were added to be able to follow the
offers partners propose, as well as the salesperson responsible for those
sales. Finally tags were added to be able to track special qualities or detail
some of the properties have that might be of interest to potential buyers.

Proper relationships were added between `estate.property` and the new
models, being them as follow

 -  `estate.property -- One2many --> estate.property.offer`
 -  `estate.property -- Many2one --> estate.property.type`
 -  `estate.property -- Many2many --> estate.property.tag`