[IMP] estate: add computed fields and onchanges T#69469
Add computed fields for `estate.property.offer` and `estate.property`
and onchange for `estate.property`.

Everything was implemented as specified in the section "Getting Started 9".

Computed fields for `estate.property.offer` were added to track the
validity of an offer and its corresponding deadline. Computed fields for
`estate.property` were added to track the `total_area` of the property
and the `best_price` from all the offers.

Onchange for `garden_area` and `garden_orientation` were added to
improve UX for filling those fields when `garden` is set or unset.