[IMP] estate: add actions to `property` and `offer` T#69469
Add buttons and actions for accepting and rejecting offers. Add buttons
and actions for setting the state of a property as sold or canceled.

Everything was implemented as specified in the section "Getting Started
10".

Actions attached to buttons help the state of `estate.property` to stay
consistent, as the user does not have to update all the corresponding
fields manually. For this multiple interactions were added to help the
user. These interactions and checks are as follow

 1. If a property is canceled then all the offers are refused.
 2. If an offer is accepted, then all other offers are refused.
 3. If an offer is accepted, then the fields `selling_price` and `buyer`
    are filled up.
 4. If a property is sold or canceled, none of the other actions will be
    allowed and an `UserError` will be thrown.