[IMP] estate: extend models and add UI for users T#69469
In this commit I extend `estate.property` CRUD methods so it cannot be
possible to delete properties that are neither in `"new"` or
`"canceled"` states.  Model `res.users` is extended to include the
properties sold by an user and show them in its form view.

All changes were made accordingly to the chapter "Getting Started 13".

These changes were added to improve UI/UX of users for the module.