[IMP] estate: add invoicing to `estate.property` T#69469
Added `estate_account` module that links `estate` and `account` modules.
An invoice is created when a property is marked as sold if the module
`estate_account` is installed.

For the invoice to be created a journal and account must be setup for
the partner buying the property and the "Invoicing" module installed.

All changes were made accordingly to the chapter "Getting Started 14".