[IMP] estate: add demo and functional data T#69469
In this commit I add functional data for `estate.property.type`, and
demo data for `estate.property`. `estate.property.type` is added to the
module, as to provide minimal functionality for the former and examples
for the later two.

Everything is implemented as specified in the "Define module data"
section in the technical training.

It was checked that the state of the instances of `estate.property` in
the demo data was consistent, and that the order of loading of demo data
was consistent with the constraints on the data.