[FIX] estate: external id's of views for all models T#69469
As part of the section "Restrict access to data" it is found the
documentation on "Coding guidelines". View records where not assigned
correct external id's accordingly to the coding guidelines, as
the used `id` convention was `module.model.view_type`. This is fixed to
changing them to use the convention `module.model.view.view_type`.

The same happened with the inherited views. The `id` of the inherited
views was using the

    `module_src.model.view_type.inherit.module_dest`

convention instead of

    `module_src.model.view.view_type.inherit.module_dest`.

XML files where not declared with the correct encoding
`encoding="utf-8"`, which is also corrected in this commit.