[IMP] estate: add unit test T#69469
In this commit, unit tests are added, following the specifications of the
section "Safeguard your code with unit test" of the technical training.

Tests were added for the `onchange` registered for `garden`, which set
and unset `garden_area` and `garden_orientation`. For the action
`action_sold` of `estate.property`, tests were added to check that no
property without an accepted offer can be sold, and not offer can be
created for a sold property.