[IMP] estate: add reports for properties and users T#69469
In this commit reports are added, following the specifications of the
section "Build PDF Reports" of the technical specification.

Reports for individual properties were added, where it shows its offers.
Reports for users where added too, allowing to see all the properties
that they can manage and its corresponding offers.

Additional demo data was added to test the creation of reports locally.