[FIX] estate: fix error message and check T#69469
When a sold property was tried to be sold again, the message showed was
`"Cannot sell a property without an accepted offer."`, which is
incorrect. The series of events that lead to this to happen were

 1.1 `_check_accepted_offer` ran before `_check_availability`,
 1.2 `_check_accepted_offer` checked that `status` was not
     `offer_accepted`, which did not cover `sold`, to raise an
     exception.

The fixes for both mistakes are

 2.1 in `action_mark_sold` move `_check_accepted_offer` _after_
     `_check_availability`,
 2.2 in `_check_accepted_offer`, verify that `status not in
     ("offer_accepted", "sold")`

If [2.1] is applied, then for [1.2] the case where `status` is `sold`
will never be reached, making [2.2] redundant. I decided to left this as
an extra check if something about `status` were to change.