[REF] estate: refactor `_compute_date_deadline` T#69469
When computing the `record.date_deadline` a branch was introduced to
account for the fact that `record.create_date` can be `False`.
Specifically it did not have a value assigned when a new record was
being created.

This refactor makes more evident that `record.create_date` is the
preferred value but, if it doesn't exist, then `context_today(record)`
will be called.